requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
psycopg2-binary==2.8.6
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
